requirements.txt
streamlit
pydantic
groq
requests
python-dotenv